complete jwt authentication

+++ backend/controllers/goalController.js
@@ -1,5 +1,6 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 // @desc Get goals
 // @route GET /api/goals
 // @access Private
@@ -16,9 +17,13 @@ const setGoals = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Please add a text field");
   }
+
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
+
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goal);
 });
 
@@ -31,6 +36,20 @@ const updateGoals = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal not found ");
   }
+  const user = await User.findById(req.user.id);
+
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("user not authorized");
+  }
+
   //(which id to be updated, which to update, {new: true}/{which field you wanna change:value})
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
@@ -47,6 +66,21 @@ const deleteGoals = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal not found ");
   }
+
+  const user = await User.findById(req.user.id);
+
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("user not authorized");
+  }
+
   await Goal.findByIdAndDelete(req.params.id);
   res.status(200).json({ id: req.params.id });
 });
